tunnel: Allow normalizing `TracingEvent`s
